Simplify S3 vector index into single struct
