+++ crates/runtime/src/embeddings/index/mod.rs
@@ -18,31 +18,15 @@ use std::{any::Any, collections::HashSet, sync::Arc};
 
 use arrow::array::RecordBatch;
 use arrow_schema::{ArrowError, Field, Fields, SchemaRef};
-use async_openai::types::EmbeddingInput;
 use async_trait::async_trait;
-use data_components::s3_vectors::{
-    MetadataColumns, S3_VECTOR_EMBEDDING_NAME, S3_VECTOR_PRIMARY_KEY_NAME,
-    list_provider::S3VectorsListTable, query_provider::S3VectorsQueryTable,
-};
+use data_components::s3_vectors::MetadataColumns;
 use llms::embeddings::Embed;
 use runtime_datafusion_index::Index;
-use search::SEARCH_SCORE_COLUMN_NAME;
 use snafu::ResultExt;
 
-use crate::{
-    embedding_col, embeddings::index::query_table::metadata_columns_to_exprs,
-    model::EmbeddingModelStore,
-};
 use datafusion::{
-    catalog::TableProvider,
-    common::Column,
-    datasource::{DefaultTableSource, ViewTable},
-    error::DataFusionError,
-    logical_expr::{BinaryExpr, Cast, LogicalPlan, Operator, Projection, TableScan, expr::Alias},
-    prelude::{Expr, lit},
-    sql::TableReference,
+    catalog::TableProvider, error::DataFusionError, logical_expr::LogicalPlan, prelude::Expr,
 };
-use tokio::sync::RwLock;
 
 pub(crate) mod query_table;
 mod retry_client;
@@ -86,193 +70,6 @@ pub trait VectorIndex: std::fmt::Debug + Send + Sync {
         &self,
         query: &str,
     ) -> Result<Arc<dyn TableProvider>, Box<dyn std::error::Error + Send + Sync>>;
-}
-
-#[derive(Debug, Clone)]
-pub struct IndexEmbeddingConfig {
-    pub model_name: String,
-    pub embedding_models: Arc<RwLock<EmbeddingModelStore>>,
-}
-
-/// Implementations of indexes that can produce embedding vectors for a column in the associated [`IndexedTableProvider`], and some, provide efficient search mechanism for it.
-#[derive(Debug, Clone)]
-pub struct S3Vector {
-    index: s3::S3VectorIndex,
-    cfg: IndexEmbeddingConfig,
-}
-
-impl S3Vector {
-    #[must_use]
-    pub fn new(index: s3::S3VectorIndex, cfg: IndexEmbeddingConfig) -> Self {
-        Self { index, cfg }
-    }
-
-    pub async fn embedding_model(&self) -> Option<Arc<dyn Embed>> {
-        let model_lock = self.cfg.embedding_models.read().await;
-        let model = model_lock.get(&self.cfg.model_name)?;
-        Some(Arc::clone(model))
-    }
-
-    pub async fn query_vector(
-        &self,
-        query: &str,
-    ) -> Result<Vec<f32>, Box<dyn std::error::Error + Send + Sync>> {
-        let models = self.cfg.embedding_models.read().await;
-        let Some(embedding_model) = models.get(&self.cfg.model_name) else {
-            return Err(Box::from(format!(
-                "Vector index requires '{}' embedding model, but is not available.",
-                self.cfg.model_name
-            )));
-        };
-        let mut resp = embedding_model
-            .embed(EmbeddingInput::String(query.to_string()))
-            .await
-            .boxed()?;
-        let Some(query_vector) = resp.pop() else {
-            return Err(Box::from(format!(
-                "Embedding model '{}' produced no embedding for the query '{query}'.",
-                self.cfg.model_name,
-            )));
-        };
-
-        Ok(query_vector)
-    }
-}
-
-#[async_trait]
-impl VectorIndex for S3Vector {
-    fn embedded_column(&self) -> String {
-        self.index.embedded_column.clone()
-    }
-
-    fn primary_fields(&self) -> Vec<Field> {
-        self.index.primary_key.clone()
-    }
-
-    /// Use a [`S3VectorsListTable`] and then:
-    ///   1. Convert the primary key to its appropriate name and data type
-    ///   2. Rename [`S3_VECTOR_EMBEDDING_NAME`] appropriately
-    fn list_table_provider(
-        &self,
-    ) -> Result<Arc<dyn TableProvider>, Box<dyn std::error::Error + Send + Sync>> {
-        let Some((pk_name, pk_data_type)) = self
-            .primary_fields()
-            .first()
-            .map(|f| (f.name().clone(), f.data_type().clone()))
-        else {
-            return Err(Box::from(
-                "Vector indexes defined without a primary key cannot be used for retrieving vectors"
-                    .to_string(),
-            ));
-        };
-
-        let mut projection = metadata_columns_to_exprs(self.metadata_columns());
-        projection.extend(vec![
-            Expr::Alias(Alias::new(
-                Expr::Cast(Cast::new(
-                    Box::new(Expr::Column(datafusion::common::Column::new_unqualified(
-                        S3_VECTOR_PRIMARY_KEY_NAME,
-                    ))),
-                    pk_data_type,
-                )),
-                None::<TableReference>,
-                pk_name,
-            )),
-            Expr::Alias(Alias::new(
-                Expr::Column(datafusion::common::Column::new_unqualified(
-                    S3_VECTOR_EMBEDDING_NAME,
-                )),
-                None::<TableReference>,
-                embedding_col!(self.embedded_column()),
-            )),
-        ]);
-
-        table_with_projection(
-            Arc::new(S3VectorsListTable::from(self.index.table.clone())),
-            projection,
-        )
-        .boxed()
-    }
-
-    fn metadata_columns(&self) -> &MetadataColumns {
-        &self.index.metadata_columns
-    }
-
-    async fn write(&self, record: &RecordBatch) {
-        s3::write(&self.index, &self.cfg, record).await;
-    }
-
-    async fn query_table_provider(
-        &self,
-        query: &str,
-    ) -> Result<Arc<dyn TableProvider>, Box<dyn std::error::Error + Send + Sync>> {
-        let Some((pk_name, pk_data_type)) = self
-            .primary_fields()
-            .first()
-            .map(|f| (f.name().clone(), f.data_type().clone()))
-        else {
-            return Err(Box::from(
-                "Vector indexes defined without a primary key cannot be used for querying vectors"
-                    .to_string(),
-            ));
-        };
-
-        let mut projection = vec![
-            Expr::Alias(Alias::new(
-                Expr::Cast(Cast::new(
-                    Box::new(Expr::Column(Column::new_unqualified(
-                        S3_VECTOR_PRIMARY_KEY_NAME,
-                    ))),
-                    pk_data_type.clone(),
-                )),
-                None::<TableReference>,
-                pk_name,
-            )),
-            Expr::Alias(Alias::new(
-                Expr::Column(Column::new_unqualified(S3_VECTOR_EMBEDDING_NAME)),
-                None::<TableReference>,
-                embedding_col!(self.embedded_column()),
-            )),
-            Expr::Alias(Alias::new(
-                Expr::BinaryExpr(BinaryExpr::new(
-                    Box::new(lit(1.0)),
-                    Operator::Minus,
-                    Box::new(Expr::Column(Column::new_unqualified("distance"))),
-                )),
-                None::<TableReference>,
-                SEARCH_SCORE_COLUMN_NAME,
-            )),
-        ];
-        projection.extend(metadata_columns_to_exprs(self.metadata_columns()));
-
-        // TODO: Restructure [`S3VectorsQueryTable`] to take an async function (probably a trait)
-        // like `async fn(&str) -> vec<f32>`, to avoid early embedding request.
-        let vector = self.query_vector(query).await?;
-        let tp = Arc::new(S3VectorsQueryTable::new(self.index.table.clone(), vector));
-
-        table_with_projection(tp, projection).boxed()
-    }
-}
-
-#[async_trait]
-impl Index for S3Vector {
-    fn name(&self) -> &'static str {
-        "s3_vector_index"
-    }
-
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-
-    fn required_columns(&self) -> Vec<String> {
-        self.index.required_columns()
-    }
-
-    async fn compute_index(&self, batches: Vec<RecordBatch>) {
-        for rb in batches {
-            self.write(&rb).await;
-        }
-    }
 }
 
 // Returns true if the vector index table has all requested columns and can handle all filters (i.e. filters pertain to vector index column, even if they must be post-applied in DataFusion).
@@ -336,26 +133,6 @@ pub(super) fn vector_index_filters(
         })
         .cloned()
         .collect()
-}
-
-fn table_with_projection(
-    tbl: Arc<dyn TableProvider>,
-    projection: Vec<Expr>,
-) -> Result<Arc<dyn TableProvider>, DataFusionError> {
-    let scan = TableScan::try_new(
-        "tbl",
-        Arc::new(DefaultTableSource::new(tbl)),
-        None,
-        vec![],
-        None,
-    )?;
-    Ok(Arc::new(ViewTable::new(
-        LogicalPlan::Projection(Projection::try_new(
-            projection,
-            Arc::new(LogicalPlan::TableScan(scan)),
-        )?),
-        None,
-    )) as Arc<dyn TableProvider>)
 }
 
 // Returns a new projection without `columns` in the projection.
